+++ test/sample_apps/minimal_phoenix14/lib/minimal_phoenix14_web/gettext.ex
@@ -1,0 +1,3 @@
+defmodule MinimalPhoenix14Web.Gettext do
+  use Gettext, otp_app: :minimal_phoenix14
+end